uskat.s5,S5SequenceBlob: using byte[] to store values of any type.

+++ org.toxsoft.uskat.s5.histdata10/src/org/toxsoft/uskat/s5/server/backend/supports/histdata/impl/sequences/async/S5HistDataAsyncFloatingBlobEntity0.java
@@ -15,7 +15,7 @@ import jakarta.persistence.*;
  */
 @Entity
 public class S5HistDataAsyncFloatingBlobEntity0
-    extends S5SequenceAsyncBlob<S5HistDataAsyncFloatingEntity0, double[], double[]> {
+    extends S5SequenceAsyncBlob<S5HistDataAsyncFloatingEntity0, double[], byte[]> {
 
   private static final long serialVersionUID = 157157L;
