+++ org.toxsoft.uskat.s5.histdata10/src/org/toxsoft/uskat/s5/server/backend/supports/histdata/impl/sequences/async/S5HistDataAsyncIntegerBlobEntity3.java
@@ -15,7 +15,7 @@ import jakarta.persistence.*;
  */
 @Entity
 public class S5HistDataAsyncIntegerBlobEntity3
-    extends S5SequenceAsyncBlob<S5HistDataAsyncIntegerEntity3, long[], long[]> {
+    extends S5SequenceAsyncBlob<S5HistDataAsyncIntegerEntity3, long[], byte[]> {
 
   private static final long serialVersionUID = 157157L;
