+++ org.toxsoft.uskat.s5.histdata10/src/org/toxsoft/uskat/s5/server/backend/supports/histdata/impl/sequences/async/S5HistDataAsyncTimestampBlobEntity8.java
@@ -15,7 +15,7 @@ import jakarta.persistence.*;
  */
 @Entity
 public class S5HistDataAsyncTimestampBlobEntity8
-    extends S5SequenceAsyncBlob<S5HistDataAsyncTimestampEntity8, long[], long[]> {
+    extends S5SequenceAsyncBlob<S5HistDataAsyncTimestampEntity8, long[], byte[]> {
 
   private static final long serialVersionUID = 157157L;
