+++ org.toxsoft.uskat.s5.histdata10/src/org/toxsoft/uskat/s5/server/backend/supports/histdata/impl/sequences/sync/S5HistDataSyncIntegerBlobEntity6.java
@@ -15,7 +15,7 @@ import jakarta.persistence.*;
  */
 @Entity
 public class S5HistDataSyncIntegerBlobEntity6
-    extends S5SequenceSyncBlob<S5HistDataSyncIntegerEntity6, long[], long[]> {
+    extends S5SequenceSyncBlob<S5HistDataSyncIntegerEntity6, long[], byte[]> {
 
   private static final long serialVersionUID = 157157L;
